+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/util/ByteBufferPoolImpl3.java
@@ -29,6 +29,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.SortedMap;
 import java.util.TreeMap;
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -250,6 +251,26 @@ public class ByteBufferPoolImpl3 implements ByteBufferPool {
         } finally {
             if (buffer != null) {
                 release(buffer);
+            }
+        }
+    }
+
+    @Override
+    public void doWithBufferPair(final int minKeyCapacity,
+                                 final int minValueCapacity,
+                                 final BiConsumer<ByteBuffer, ByteBuffer> work) {
+        ByteBuffer keyBuffer = null;
+        ByteBuffer valueBuffer = null;
+        try {
+            keyBuffer = getBuffer(minKeyCapacity);
+            valueBuffer = getBuffer(minValueCapacity);
+            work.accept(keyBuffer, valueBuffer);
+        } finally {
+            if (keyBuffer != null) {
+                release(keyBuffer);
+            }
+            if (valueBuffer != null) {
+                release(valueBuffer);
             }
         }
     }