+++ stroom-pipeline/src/test/java/stroom/pipeline/refdata/store/offheapstore/databases/TestValueStoreDb.java
@@ -1,9 +1,12 @@
 package stroom.pipeline.refdata.store.offheapstore.databases;
 
 
+import stroom.pipeline.refdata.store.BasicValueStoreHashAlgorithmImpl;
 import stroom.pipeline.refdata.store.ByteBufferPoolFactory;
 import stroom.pipeline.refdata.store.RefDataValue;
 import stroom.pipeline.refdata.store.StringValue;
+import stroom.pipeline.refdata.store.ValueStoreHashAlgorithm;
+import stroom.pipeline.refdata.store.XxHashValueStoreHashAlgorithm;
 import stroom.pipeline.refdata.store.offheapstore.ValueStoreKey;
 import stroom.pipeline.refdata.store.offheapstore.lmdb.EntryConsumer;
 import stroom.pipeline.refdata.store.offheapstore.lmdb.LmdbUtils;
@@ -12,8 +15,12 @@ import stroom.pipeline.refdata.store.offheapstore.serdes.RefDataValueSerdeFactor
 import stroom.pipeline.refdata.store.offheapstore.serdes.ValueStoreKeySerde;
 import stroom.pipeline.refdata.util.PooledByteBuffer;
 
+import io.vavr.Tuple;
+import io.vavr.Tuple2;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.lmdbjava.KeyRange;
 import org.lmdbjava.Txn;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,19 +38,31 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class TestValueStoreDb extends AbstractLmdbDbTest {
     private static final Logger LOGGER = LoggerFactory.getLogger(TestValueStoreDb.class);
+
     private final RefDataValueSerdeFactory refDataValueSerdeFactory = new RefDataValueSerdeFactory();
+    private final ValueStoreHashAlgorithm xxHashAlgorithm = new XxHashValueStoreHashAlgorithm();
+    private final ValueStoreHashAlgorithm basicHashAlgorithm = new BasicValueStoreHashAlgorithmImpl();
     private ValueStoreDb valueStoreDb = null;
 
-    //TODO should really spin up guice rather than use this factory class
-//    private final RefDataValueSerde refDataValueSerde = RefDataValueSerdeFactory.create();
 
     @BeforeEach
     void setup() {
+        // the default
         valueStoreDb = new ValueStoreDb(
                 lmdbEnv,
                 new ByteBufferPoolFactory().getByteBufferPool(),
                 new ValueStoreKeySerde(),
-                new GenericRefDataValueSerde(refDataValueSerdeFactory));
+                new GenericRefDataValueSerde(refDataValueSerdeFactory),
+                xxHashAlgorithm);
+    }
+
+    private void setupValueStoreDb(final ValueStoreHashAlgorithm valueStoreHashAlgorithm) {
+        valueStoreDb = new ValueStoreDb(
+                lmdbEnv,
+                new ByteBufferPoolFactory().getByteBufferPool(),
+                new ValueStoreKeySerde(),
+                new GenericRefDataValueSerde(refDataValueSerdeFactory),
+                valueStoreHashAlgorithm);
     }
 
 
@@ -63,7 +82,11 @@ class TestValueStoreDb extends AbstractLmdbDbTest {
 
     @Test
     void testGetOrCreateSparseIds() {
-        final List<String> stringsWithSameHash = generateHashClashes(10);
+        // We have to set up the DB with the basic hash func so we can be assured of hash clashes
+        setupValueStoreDb(basicHashAlgorithm);
+
+        final List<String> stringsWithSameHash = generateHashClashes(
+                10, valueStoreDb.getValueStoreHashAlgorithm());
 
         final List<RefDataValue> refDataValues = stringsWithSameHash.stream()
                 .map(StringValue::of)
@@ -117,11 +140,74 @@ class TestValueStoreDb extends AbstractLmdbDbTest {
         assertThat(valueStoreKeysMap.get(9).getUniqueId()).isEqualTo((short) 6);
     }
 
+    @Test
+    void testGetOrCreateSparseIds2() {
+        // We have to set up the DB with the basic hash func so we can be assured of hash clashes
+        setupValueStoreDb(basicHashAlgorithm);
+
+        final List<String> stringsWithSameHash = generateHashClashes(
+                10, valueStoreDb.getValueStoreHashAlgorithm());
+
+        final List<RefDataValue> refDataValues = stringsWithSameHash.stream()
+                .map(StringValue::of)
+                .collect(Collectors.toList());
+
+        final Map<Integer, ValueStoreKey> valueStoreKeysMap = new HashMap<>();
+
+        LmdbUtils.doWithWriteTxn(lmdbEnv, writeTxn -> {
+            ValueStoreKey valueStoreKey;
+
+            // insert the first five values, all have same hash so should get increasing
+            // id values.
+            for (int i = 0; i < 5; i++) {
+                valueStoreKey = getOrCreate(writeTxn, refDataValues.get(i));
+                valueStoreKeysMap.put(i, valueStoreKey);
+                LOGGER.info("Assigned id: {}", valueStoreKey.getUniqueId());
+                assertThat(valueStoreKey.getUniqueId()).isEqualTo((short) i);
+            }
+        });
+        assertThat(valueStoreDb.getEntryCount()).isEqualTo(5);
+        valueStoreDb.logDatabaseContents();
+
+        // delete values 0,1,2 leaving ids 3,4
+        for (int i = 0; i < 3; i++) {
+            valueStoreDb.delete(valueStoreKeysMap.get(i));
+        }
+        assertThat(valueStoreDb.getEntryCount()).isEqualTo(2);
+        valueStoreDb.logDatabaseContents();
+
+        // now put the remaining 5 values
+        // they should fill up the gaps in the ID sequence
+        LmdbUtils.doWithWriteTxn(lmdbEnv, writeTxn -> {
+            ValueStoreKey valueStoreKey;
+            for (int i = 5; i < 10; i++) {
+                valueStoreKey = getOrCreate(writeTxn, refDataValues.get(i));
+                valueStoreKeysMap.put(i, valueStoreKey);
+                LOGGER.info("Assigned id: {}", valueStoreKey.getUniqueId());
+            }
+        });
+        assertThat(valueStoreDb.getEntryCount()).isEqualTo(5 - 3 + 5);
+        valueStoreDb.logDatabaseContents();
+
+        // check the ID value for each of our original values in insertion order
+        assertThat(valueStoreKeysMap.get(0).getUniqueId()).isEqualTo((short) 0); //since deleted
+        assertThat(valueStoreKeysMap.get(1).getUniqueId()).isEqualTo((short) 1); //since deleted
+        assertThat(valueStoreKeysMap.get(2).getUniqueId()).isEqualTo((short) 2); //since deleted
+        assertThat(valueStoreKeysMap.get(3).getUniqueId()).isEqualTo((short) 3);
+        assertThat(valueStoreKeysMap.get(4).getUniqueId()).isEqualTo((short) 4);
+        assertThat(valueStoreKeysMap.get(5).getUniqueId()).isEqualTo((short) 0); //used empty space after delete
+        assertThat(valueStoreKeysMap.get(6).getUniqueId()).isEqualTo((short) 1); //used empty space after delete
+        assertThat(valueStoreKeysMap.get(7).getUniqueId()).isEqualTo((short) 2); //used empty space after delete
+        assertThat(valueStoreKeysMap.get(8).getUniqueId()).isEqualTo((short) 5);
+        assertThat(valueStoreKeysMap.get(9).getUniqueId()).isEqualTo((short) 6);
+    }
+
     /**
      * Generate a list (of size desiredRecords) of strings that all share the same hashcode
      * Adapted from https://gist.github.com/vaskoz/5703423
      */
-    public List<String> generateHashClashes(final int desiredRecords) {
+    public List<String> generateHashClashes(final int desiredRecords,
+                                            final ValueStoreHashAlgorithm valueStoreHashAlgorithm) {
 
         List<String> strings = new ArrayList<>(Arrays.asList("Aa", "BB"));
         List<String> temp = new ArrayList<>();
@@ -146,7 +232,10 @@ class TestValueStoreDb extends AbstractLmdbDbTest {
 
         //dbl check all have same hash
         assertThat(strings.stream()
-                .map(String::hashCode)
+                .peek(str -> {
+                    LOGGER.info("{} {} {}", str, str.hashCode(), valueStoreHashAlgorithm.hash(str));
+                })
+                .map(valueStoreHashAlgorithm::hash)
                 .distinct()
                 .count()).isEqualTo(1);
 
@@ -178,24 +267,33 @@ class TestValueStoreDb extends AbstractLmdbDbTest {
         LmdbUtils.doWithReadTxn(lmdbEnv, txn -> {
             // lookup our three entries
             valueToKeyMap.forEach((valueStr, valueStoreKey) -> {
-                ByteBuffer valueBuffer = valueStoreDb.getAsBytes(txn, keySerde.serialize(valueStoreKey)).get();
-                RefDataValue val = refDataValueSerdeFactory.deserialize(valueBuffer, StringValue.TYPE_ID);
-                assertThat(val).isInstanceOf(StringValue.class);
-                assertThat(((StringValue) val).getValue()).isEqualTo(valueStr);
+                valueStoreDb.getPooledKeyBuffer().doWithByteBuffer(keyBuffer -> {
+                    keySerde.serialize(keyBuffer, valueStoreKey);
+                    ByteBuffer valueBuffer = valueStoreDb.getAsBytes(txn, keyBuffer).get();
+                    RefDataValue val = refDataValueSerdeFactory.deserialize(valueBuffer, StringValue.TYPE_ID);
+                    assertThat(val).isInstanceOf(StringValue.class);
+                    assertThat(((StringValue) val).getValue()).isEqualTo(valueStr);
+                });
             });
 
             // now try and get a value that doesn't exist
-            Optional<RefDataValue> optRefDataValue = valueStoreDb.get(
-                    txn,
-                    keySerde.serialize(new ValueStoreKey(123456, (short) 99)),
-                    StringValue.TYPE_ID);
+            valueStoreDb.getPooledKeyBuffer().doWithByteBuffer(keyBuffer -> {
+                keySerde.serialize(keyBuffer, new ValueStoreKey(123456, (short) 99));
+                Optional<RefDataValue> optRefDataValue = valueStoreDb.get(
+                        txn,
+                        keyBuffer,
+                        StringValue.TYPE_ID);
+                assertThat(optRefDataValue).isEmpty();
+            });
 
-            assertThat(optRefDataValue).isEmpty();
         });
     }
 
     @Test
     void testGet_sameHashCodes() {
+        // We have to set up the DB with the basic hash func so we can be assured of hash clashes
+        setupValueStoreDb(basicHashAlgorithm);
+
         assertThat(valueStoreDb.getEntryCount()).isEqualTo(0);
 
         String val1str = "AaAa";
@@ -231,11 +329,14 @@ class TestValueStoreDb extends AbstractLmdbDbTest {
         LmdbUtils.doWithReadTxn(lmdbEnv, txn -> {
             // lookup our entries
             List<Integer> ids = new ArrayList<>();
-            valueToKeyMap.forEach((valueStr, valueStoreKey) -> {
-                ByteBuffer valueBuffer = valueStoreDb.getAsBytes(txn, keySerde.serialize(valueStoreKey)).get();
-                RefDataValue val = refDataValueSerdeFactory.deserialize(valueBuffer, StringValue.TYPE_ID);
-                assertThat(val).isInstanceOf(StringValue.class);
-                assertThat(((StringValue) val).getValue()).isEqualTo(valueStr);
+            valueStoreDb.getPooledKeyBuffer().doWithByteBuffer(keyBuffer -> {
+                valueToKeyMap.forEach((valueStr, valueStoreKey) -> {
+                    keySerde.serialize(keyBuffer, valueStoreKey);
+                    ByteBuffer valueBuffer = valueStoreDb.getAsBytes(txn, keyBuffer).get();
+                    RefDataValue val = refDataValueSerdeFactory.deserialize(valueBuffer, StringValue.TYPE_ID);
+                    assertThat(val).isInstanceOf(StringValue.class);
+                    assertThat(((StringValue) val).getValue()).isEqualTo(valueStr);
+                });
             });
         });
     }
@@ -261,6 +362,37 @@ class TestValueStoreDb extends AbstractLmdbDbTest {
             boolean areValuesEqual = valueStoreDb.areValuesEqual(writeTxn, unknownValueStoreKeyBuffer, value2);
             assertThat(areValuesEqual).isFalse();
         });
+    }
+
+    @Test
+    void testKeyOrder() {
+
+        // Ensure entries come back in the right order
+        long hash = 123456789;
+        final List<Tuple2<ValueStoreKey, RefDataValue>> data = List.of(
+                Tuple.of(new ValueStoreKey(hash, (short) 0), new StringValue("val1")),
+                Tuple.of(new ValueStoreKey(hash, (short) 1), new StringValue("val2")),
+                Tuple.of(new ValueStoreKey(hash, (short) 2), new StringValue("val3")),
+                Tuple.of(new ValueStoreKey(hash, (short) 3), new StringValue("val4")));
+
+        LmdbUtils.doWithWriteTxn(lmdbEnv, writeTxn -> {
+            data.forEach(tuple -> {
+                valueStoreDb.put(writeTxn, tuple._1(), tuple._2(), false);
+            });
+        });
+
+        final KeyRange<ValueStoreKey> keyRangeAll = KeyRange.all();
+
+        final List<ValueStoreKey> output = LmdbUtils.getWithReadTxn(lmdbEnv, readTxn ->
+                valueStoreDb.streamEntries(readTxn, keyRangeAll, stream ->
+                        stream
+                                .map(Tuple2::_1)
+                                .collect(Collectors.toList())));
+
+        Assertions.assertThat(output)
+                .containsExactlyElementsOf(data.stream()
+                        .map(Tuple2::_1)
+                        .collect(Collectors.toList()));
     }
 
     private void doAreValuesEqualAssert(final StringValue value1, final StringValue value2, final boolean expectedResult) {