+++ stroom-pipeline/src/test/java/stroom/pipeline/refdata/store/offheapstore/serdes/TestRangeStoreKeySerde.java
@@ -18,16 +18,19 @@
 package stroom.pipeline.refdata.store.offheapstore.serdes;
 
 
-import org.junit.jupiter.api.Test;
 import stroom.pipeline.refdata.store.offheapstore.RangeStoreKey;
 import stroom.pipeline.refdata.store.offheapstore.UID;
 import stroom.util.shared.Range;
+
+import org.junit.jupiter.api.Test;
+
+import java.nio.ByteBuffer;
 
 class TestRangeStoreKeySerde extends AbstractSerdeTest<RangeStoreKey, RangeStoreKeySerde> {
 
     @Test
     void testSerialiseDeserialise() {
-        final UID uid = UID.of(0, 1, 2, 3);
+        final UID uid = UID.of(ByteBuffer.allocateDirect(UID.UID_ARRAY_LENGTH), 0, 1, 2, 3);
         final Range<Long> range = new Range<>(23L, 52L);
 
         final RangeStoreKey rangeStoreKey = new RangeStoreKey(uid, range);