+++ stroom-pipeline/src/test/java/stroom/pipeline/refdata/store/offheapstore/serdes/TestValueStoreKeySerde.java
@@ -1,8 +1,9 @@
 package stroom.pipeline.refdata.store.offheapstore.serdes;
 
 
-import org.junit.jupiter.api.Test;
 import stroom.pipeline.refdata.store.offheapstore.ValueStoreKey;
+
+import org.junit.jupiter.api.Test;
 
 import java.nio.ByteBuffer;
 
@@ -13,30 +14,15 @@ class TestValueStoreKeySerde extends AbstractSerdeTest<ValueStoreKey, ValueStore
     @Test
     void testSerializeDeserialize() {
         final ValueStoreKey valueStoreKey = new ValueStoreKey(
-                123456789,
+                123456789L,
                 (short) 1);
 
         doSerialisationDeserialisationTest(valueStoreKey);
     }
 
     @Test
-    void testNextId() {
-
-        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567, (short) 123);
-
-        ByteBuffer originalBuffer = serialize(originalValueStoreKey);
-
-        ByteBuffer mutatedBuffer = ValueStoreKeySerde.nextId(originalBuffer);
-
-        ValueStoreKey newValueStoreKey = deserialize(mutatedBuffer);
-
-        assertThat(newValueStoreKey.getValueHashCode()).isEqualTo(originalValueStoreKey.getValueHashCode());
-        assertThat(newValueStoreKey.getUniqueId()).isEqualTo((short) (originalValueStoreKey.getUniqueId() + 1));
-    }
-
-    @Test
     void testIncrementId() {
-        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567, (short) 123);
+        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567L, (short) 123);
 
         ByteBuffer byteBuffer = serialize(originalValueStoreKey);
 
@@ -50,11 +36,13 @@ class TestValueStoreKeySerde extends AbstractSerdeTest<ValueStoreKey, ValueStore
 
     @Test
     void testUpdateId() {
-        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567, (short) 123);
+        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567L, (short) 123);
 
         ValueStoreKeySerde serde = new ValueStoreKeySerde();
 
-        ByteBuffer byteBuffer = serde.serialize(originalValueStoreKey);
+        ByteBuffer byteBuffer = ByteBuffer.allocateDirect(serde.getBufferCapacity());
+
+        serde.serialize(byteBuffer, originalValueStoreKey);
 
         ValueStoreKeySerde.updateId(byteBuffer, (short) 456);
 
@@ -67,11 +55,12 @@ class TestValueStoreKeySerde extends AbstractSerdeTest<ValueStoreKey, ValueStore
     @Test
     void testExtractId() {
         short id = 123;
-        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567, id);
+        ValueStoreKey originalValueStoreKey = new ValueStoreKey(1234567L, id);
 
         ValueStoreKeySerde serde = new ValueStoreKeySerde();
 
-        ByteBuffer byteBuffer = serde.serialize(originalValueStoreKey);
+        ByteBuffer byteBuffer = ByteBuffer.allocateDirect(serde.getBufferCapacity());
+        serde.serialize(byteBuffer, originalValueStoreKey);
 
         short extractedId = ValueStoreKeySerde.extractId(byteBuffer);